segmentation/requirements_seg.txt
torch
torchvision
numpy
pillow
scikit-learn
tqdm
matplotlib
visdom

